arc: Add support to cache extracted raw videos in encode tests.
As the raw video files used to test encoding are too large to download
and store on a DUT, videos are stored in the vp9.webm format. Before
each test these videos are extracted to create the raw input video.
This process can take a lot of time, so this CL adds support to cache
the extracted raw video file. This saves about 30s test time when
running 1080p encode testcases. The feature is disabled by default as
we don't want to have DUT's run out of memory in the lab.

BUG=None
TEST=various video encode tests on Hatch

+++ src/chromiumos/tast/local/bundles/cros/arc/video/accel_video.go
@@ -60,7 +60,8 @@ const (
 // It pushes the binary files with different ABI and testing video data into ARC, and runs each binary for each binArgs.
 // pv is optional value, passed when we run performance test and record measurement value.
 // Note: pv must be provided when measureUsage is set at binArgs.
-func runARCVideoEncoderTest(ctx context.Context, s *testing.State, a *arc.ARC, opts encoding.TestOptions, pv *perf.Values, bas ...binArgs) {
+func runARCVideoEncoderTest(ctx context.Context, s *testing.State, a *arc.ARC,
+	opts encoding.TestOptions, cacheExtractedVideo bool, pv *perf.Values, bas ...binArgs) {
 	// Install the test APK and grant permissions
 	apkName, err := c2e2etest.ApkNameForArch(ctx, a)
 	if err != nil {
@@ -79,11 +80,20 @@ func runARCVideoEncoderTest(ctx context.Context, s *testing.State, a *arc.ARC, o
 
 	// Prepare video stream.
 	params := opts.Params
-	streamPath, err := encoding.PrepareYUV(ctx, s.DataPath(params.Name), opts.PixelFormat, params.Size)
+	streamPath := strings.TrimSuffix(s.DataPath(params.Name), ".vp9.webm")
+	switch opts.PixelFormat {
+	case videotype.I420:
+		streamPath += ".i420.yuv"
+	case videotype.NV12:
+		streamPath += ".nv12.yuv"
+	}
+	err = encoding.PrepareYUV(ctx, s.DataPath(params.Name), streamPath, opts.PixelFormat, params.Size)
 	if err != nil {
 		s.Fatal("Failed to prepare YUV file: ", err)
 	}
-	defer os.Remove(streamPath)
+	if !cacheExtractedVideo {
+		defer os.Remove(streamPath)
+	}
 
 	// Push video stream file to ARC container.
 	if err := a.Command(ctx, "mkdir", arcFilePath).Run(testexec.DumpLogOnError); err != nil {
@@ -210,18 +220,21 @@ func runARCBinaryWithArgs(ctx context.Context, s *testing.State, a *arc.ARC, com
 }
 
 // RunARCVideoTest runs all non-perf tests of arcvideoencoder_test in ARC.
-func RunARCVideoTest(ctx context.Context, s *testing.State, a *arc.ARC, opts encoding.TestOptions) {
+func RunARCVideoTest(ctx context.Context, s *testing.State, a *arc.ARC,
+	opts encoding.TestOptions, cacheExtractedVideo bool) {
 	vl, err := logging.NewVideoLogger()
 	if err != nil {
 		s.Fatal("Failed to set values for verbose logging: ", err)
 	}
 	defer vl.Close()
 
-	runARCVideoEncoderTest(ctx, s, a, opts, nil, binArgs{testFilter: "C2VideoEncoderE2ETest.Test*"})
+	//runARCVideoEncoderTest(ctx, s, a, opts, nil, binArgs{testFilter: "C2VideoEncoderE2ETest.Test*"})
+	runARCVideoEncoderTest(ctx, s, a, opts, cacheExtractedVideo, nil, binArgs{testFilter: "C2VideoEncoderE2ETest.TestSimpleEncode"})
 }
 
 // RunARCPerfVideoTest runs all perf tests of arcvideoencoder_test in ARC.
-func RunARCPerfVideoTest(ctx context.Context, s *testing.State, a *arc.ARC, opts encoding.TestOptions) {
+func RunARCPerfVideoTest(ctx context.Context, s *testing.State, a *arc.ARC,
+	opts encoding.TestOptions, cacheExtractedVideo bool) {
 	const (
 		// duration of the interval during which CPU usage will be measured.
 		measureDuration = 10 * time.Second
@@ -244,7 +257,7 @@ func RunARCPerfVideoTest(ctx context.Context, s *testing.State, a *arc.ARC, opts
 	}
 
 	pv := perf.NewValues()
-	runARCVideoEncoderTest(ctx, s, a, opts, pv,
+	runARCVideoEncoderTest(ctx, s, a, opts, cacheExtractedVideo, pv,
 		// Measure FPS and latency.
 		binArgs{
 			testFilter: "C2VideoEncoderE2ETest.Perf*",
